Make combinators polymorphic in envelope
Introduces variants with `'` for the combinators `Throws`, `Throwing`, `NoThrow`
and the `VerbWithErr` types.
These variants accept an `envel` type of kind `[*] -> * -> *` which is the error
envelope to be used.
The non-primed combinators are aliases for those, called with `Envelope`.

A typeclass `EnvelopeStatus` is introduced, allowing retrieval of HTTP status
codes from the envelope.